pyproject.toml
[tool.poetry]
name = "ottoneu"
version = "0.1.0"
description = ""
authors = ["Randy Akerman <randy@akerman.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.2"
selenium = "^4.7.2"
jinja2 = "^3.1.2"
python-slugify = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
